test(device-authenticity): extract basic data for tests
