+++ packages/device-authenticity/mocks/mockDeviceAuthenticityData.ts
@@ -1,50 +1,47 @@
-import { bufferUtils } from '@trezor/utils';
-
-import type { DeviceAuthenticityBlacklistConfig } from '../config/deviceAuthenticityBlacklistConfigTypes';
+import type { DeviceAuthenticityBlacklistConfig } from '../src/config/deviceAuthenticityBlacklistConfigTypes';
 import {
     type CertPubKeys,
     type DeviceAuthenticityConfig,
-} from '../config/deviceAuthenticityConfigTypes';
-import { type VerifyAuthenticityProofParams } from '../types';
-import {
-    matchRootPubKeyToCertificate,
-    prepareDeviceAuthenticityData,
-    verifyAuthenticityProof,
-} from '../verifyAuthenticityProof';
-import { parseCertificate } from '../x509certificate';
+} from '../src/config/deviceAuthenticityConfigTypes';
+import { type VerifyAuthenticityProofParams } from '../src/types';
+import { prepareDeviceAuthenticityData } from '../src/verifyAuthenticityProof';
 
-const CA_CERT_OPTIGA =
+// Cryptographic challenge for Trezor, used as `signedData` for `verifyAuthenticityProof`.
+// All signatures will be checked against this data.
+export const CHALLENGE = '29d0be0f3cb191c80d108359c64d22984a77ad8b99433814be31db0b6e9e7920';
+
+export const CA_CERT_OPTIGA =
     '308201df30820184a00302010202040a001ab3300a06082a8648ce3d0403023054310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3125302306035504030c1c5472657a6f72204d616e75666163747572696e6720526f6f742043413020170d3233303130313030303030305a180f32303533303130313030303030305a304f310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3120301e06035504030c175472657a6f72204d616e75666163747572696e672043413059301306072a8648ce3d020106082a8648ce3d030107034200041b36cc98d5e3d1a20677aaf26254ef3756f27c9d63080c93ad3e7d39d3ad23bf00497b924789bc8e3f87834994e16780ad4eae7e75db1f03835ca64363e980b4a3473045300e0603551d0f0101ff04040302020430120603551d130101ff040830060101ff020100301f0603551d2304183016801428b202f8f9c78a74e8c152bbfb433d99d0ca03ef300a06082a8648ce3d0403020349003046022100dfe2f837f3644c1f0250d37cd0f7d1e4e9b8cfc4820d7f5a623a8cb69df99f6c02210089148848c5fc597df4b8545d9b19d1cc15abe0e1252fa2938a4cf01ae835c563';
-const DEVICE_CERT_OPTIGA =
+export const DEVICE_CERT_OPTIGA =
     '3082019e30820145a00302010202044ee2a50f300a06082a8648ce3d040302304f310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3120301e06035504030c175472657a6f72204d616e75666163747572696e67204341301e170d3232303433303134313630315a170d3432303433303134313630315a301d311b301906035504030c1254324231205472657a6f72205361666520333059301306072a8648ce3d020106082a8648ce3d030107034200049bbf06dad9ab5905e05471ce16d5222c89c2caa39f26267ac0747129885fbd441bcc7fa84de120a36755daf30a6f47e8c0d4bddc15036ed2a3447dfa7a1d3e88a341303f300e0603551d0f0101ff040403020080300c0603551d130101ff04023000301f0603551d23041830168014176d8b9a403574f6a2b9ac353ef578682201a21a300a06082a8648ce3d04030203470030440220747c545e112df816173d3071f1ab25d399d8108550764ce1a3a428f1f18b506902200cda822c75b3da6e44e098014452f3fc324f29a79204c3fb4d5815afafc04b17';
-const SIGNATURE_OPTIGA =
+export const SIGNATURE_OPTIGA =
     '3045022100c01793ffbe4f16d4efc84a4533d9bbfbbf1baa5349346678e07fdb6d848cca7902200df11b9d2850173d9c93993fca983c6d2a3f31ea69a0e19b69e18cc3b78424fe';
 
-const SIGNATURE_OPTIGA_EVOLU =
+export const SIGNATURE_OPTIGA_EVOLU =
     '304402203e258ff50550cfb6e9634338cfd0ddc2cd5e91b3b8ae1cb9f4d0f241c40e81a8022078b2b77a527d22cc7d4a6e66301f5ed675ab285d8127095f24bd734e11484a25';
 
-const CA_CERT_TROPIC =
+export const CA_CERT_TROPIC =
     '308201c130820173a003020102020868b1982d8b917275300506032b65703054310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3125302306035504030c1c5472657a6f72204d616e75666163747572696e6720526f6f742043413020170d3235303832393132353934375a180f32303535303832323132353934375a304f310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3120301e06035504030c175472657a6f72204d616e75666163747572696e67204341302a300506032b65700321009603b4971f811ed2a1cdb9ec3e6d6d0e22facfd83892a30480460872a2003f45a3663064300e0603551d0f0101ff04040302020430120603551d130101ff040830060101ff020100301d0603551d0e04160414cf35aa12a033c044ebc6c3c0c3aefea5ae5e7db4301f0603551d2304183016801424e601ebe264f0b1cfc5bc27cc03cdf69f23a85e300506032b657003410072e527330a4b079f8b8f261489595d7e0c6bc84cd4eecdf98988f4b185df503ac3ee7364ce7ef934f56c8e823f1ac38667f85d38469884f934290efdd27bef0e';
-const DEVICE_CERT_TROPIC =
+export const DEVICE_CERT_TROPIC =
     '308201993082014ba003020102020868b1982dacb2b041300506032b6570304f310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3120301e06035504030c175472657a6f72204d616e75666163747572696e672043413020170d3235303832393132353934375a180f32303535303832323132353934375a3051311b301906035504030c1254335731205472657a6f72205361666520373121301f06035504051318343732303930323235323232323232323232323232323232310f300d060355042e130654726f706963302a300506032b6570032100c92482676994f4e2dec7eb2bb6a9ce9a8a38d4536c5e16b26646015cd7034593a341303f300e0603551d0f0101ff040403020080300c0603551d130101ff04023000301f0603551d23041830168014cf35aa12a033c044ebc6c3c0c3aefea5ae5e7db4300506032b65700341003b0aa3ccdddfd2a473e3f36059d0da2e54428c8dca5050a5fabf845ca71365d8045cc4d15c4e6e37565cbd66cdb4ba8d7ca113941f596a9bbfc25c069cf91903';
-const SIGNATURE_TROPIC =
+export const SIGNATURE_TROPIC =
     '9d6a7cfc1d9957a7eaa09f58ab385a4722dc621c6a58f0e281305f36c1447206ea6642f4e4ff36207bd57c3719101855e9c9a00a5db60cc84e20181d69f4fa00';
-const CHALLENGE = '29d0be0f3cb191c80d108359c64d22984a77ad8b99433814be31db0b6e9e7920';
 
-// TODO temporary test fixtures: when firmware is ready, regenerate with the same challenge
-const DEVICE_CERT_MCU =
+// TODO temporary test fixtures: when firmware is ready, regenerate with the same CHALLENGE instead of this one
+export const CHALLENGE_MCU = '7fd384fa2d06a8bcfd796b2b05a9c7d03a545199e8ed30c30e19d559a71796e4';
+export const DEVICE_CERT_MCU =
     '30820fdf30820655a003020102020868b1982d44540915300b06096086480165030403113054310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3125302306035504030c1c5472657a6f72204d616e75666163747572696e6720526f6f742043413020170d3235303832393132353934375a180f32303535303832323132353934375a3046311b301906035504030c1254335731205472657a6f7220536166652037311730150603550405130e3437333032313932363030303972310e300c060355042e130553544d333230820532300b06096086480165030403110382052100f71c837210b55caed1d1a401f232931ee892808bb4bbd186175e54b0fffe27a15494e5b4f548194e7c099689df4c8b207bbaa6e8e26e457581988d09fb85f5896ac13cc2e5832963099ab5dc844cdd5d3515abe7c9a2253aec5e6e98be9f5e1b8000b5c5cc52a3985ef8e0b94525811d8e31f8ea8f20d359d0b81e174a7da937a129606583f1321bcdb5067c45af60e13dd5d32ec8a328ed724882dff570c11a7bda15d8cd5995c25261887cb64c830922618df5be0d5bbd5b8a93af1baf677299a6abb986f1e88792fba0a22b223729e1a05cc55a53eb4d88ffe111aaa7cddc97764ac9438133e52528ebe622aab63ef921fa689cfbc5841fdfd8c6c8aaaf52cf55f0d6ce13e26e593bec6b57a1ef9133a6d98fbfaf630abb91a58ec27db450f31339d4ec1d3e9c1870ac8b7abaff117b41fa5adebd226e1f5326d1c06920664d71b633bef50542eee4e518ba478abafc4d539b1c9799183bc83b95bf5e0ad1df9dbd7160c053aad0ce825e5615083c0cca262a552d0b1ecb3e08d46b5364403beb1182155c1421f79bd6c53a55869bc81928b93e4902b9546fa76e47d7cb7ccb11321788e10cd354b2c0a9f4e85cfd0cbcd3827bda7e2ce9930906af9cb0ea43612847183fbcf0fea73109fbf164ae3257e02119d53614e0e4257c975c6fb9e247293b1abad833c8e16bded2f151d2dfd5e98cd15fcb930836b66a74044223715bc7e540785bd45bbbae59c7a9520ce03747f4dcc5faeeec5f99c21f5b6aa0b985e7f0690e4c58c15b32003d8a2d4c0075e1689b814e0855ab418e2853ef28f8e7e4619ad425c720ff7315d2a311761bab30c8ac9708a539b35349559b946310a7889fb479e451f396125d48fb246346aff8577e26d1ea58841ca35a6dbac02cf568d5395c399aeb44b18c6e87ceb7811bf34619e90585bbcf64521a4fd2359c5fd923b9bffc3fd423d8ce0712bcc7d7747f6d6542125449afa75adc154a758578f1cfcd8751fdf2a18c516eed70fe0b196d7917b2934dce7a3e67d204bbb4d1430824aa5e803675319715f58367f2ac7e829bc314f1f87ccd78f5e226d66d2d8f031c60fb592323d10282ad24673f6498d643f0d5fb348e00ac94d3352a6c015df8c38af4d350e4913020171994637f4486ad8ac9928582142c983c4653da7c4a2275a3083ce9cc18fef70890eb6ee91067a18722aacbbb50b43050f7fdf1721ade8c0908d05686c7db0a7c2e63c997afbb85c34eb8a0666e382c635b49c301b12887fe96a9d266ff59be7d75ef43e2a583fc012034fa32e06241bd1d24a3f8a9ddeecacba68fd679b72c50c46d9ce1f757e83f94cfbc49069bc4137bf2384ad5c2c5cdea947e64c47cd5382a0ec9e05f102152b4d8ee4fae971df142ee29ee0b094866ca7e5c4232f63e1683d6406ad43178d9ae0360c8ef43d690ab6776f04efbf9704b4e51c228ba4dfa7d58db1b9d5a1526559c4c6c0531e9fc3531718142df0de39539f1f45bc60df753e8ccddab5c2874b4f1124869c21a7701f92fcec0b3d56d79713e934f6617fb2295b29a2bc3f90e20346547ecbe1118a8f86580e310ea6b40e7445c12e1ad7c2ad11cdf090e7d00fd49c56a8b23f5bd3d01a2a8f4b30d619d57dc2158b3a2ef0ba51878e414c7f1da5711aba389fa9944ae6f1012f24a8e9b0340cc1373339a05c438983660a1525fcd93a60011df2d3749e2e19f0096bc0af765c5bf350bb414ce6797888f470624444f1d69eb8d8d4f8f556e3d3e1f27a3325633e3560c489f9aad22c50c09b7f759bbcf44eaef83389adead48884e2a9d865a54fb21d75f2c2a6fb67a7231d69a75e1ab7e16f11fe91e26a341303f300e0603551d0f0101ff040403020080300c0603551d130101ff04023000301f0603551d230418301680140bc62f01e47558497cea38acf5d4e8a8b673306a300b06096086480165030403110382097500911e8268e98e6ca111577c697ac0938864b0480c0045e01838033624ad2956c40917ec55b992d7c0ebc2945b37a4338cd7bea024679281d14f03760ab7fd66fdf3b5160f4e3164e6a19fdd414e72ee2ff01d6a4ebd51dda6321bc12f3e6d3f3d6082b1c94db2a73aab60857ea8cc97f825abcc8fc98dbb053631bdc7fd028e5bc4639ac5277b1415491b4f77830af6916f312646f25973dba75e8d77c06d9c71c5a4bf31c3c5697213fb396512e77316ab0ba7018eda7f140a1ae5c5826e263a581e1eda33a6ab55a57252f7088c12a47f5396df72f1f6911d9ac76cb7cb7e0a0f0015b0f49032015a14293a2a383e475af8348578dcf82868afdb85942bf5d7d8c349aa36691f95ab1b7fed97b9fc14a7bbc2c2d802b66e12236f221f3d39b57f9cc73285663ee2302fad3b012372d64ef83f9bec4ff1e8269d27f024f56ac1bd86dd11bb76fff9c9fae9944fbefe20fa8ad307855bf8d5f8e07d389de3bbdf0693f57df7616ca6f5b2037b20f03f09706b20c87fc8d2cc344f42ce1cef6d757640f8d1c45e7e26a66af4ca29f3f37003e2926cf6216acabb01b58d367f3e7f1bd409c0bc21d2e4875bc67608039db6fb314f990cb5115301994fec65f92cea6090adf930d7754b81451db6d85bb98b9f09e72d435d58d05a3b941c7069acc55ff8a4670696c791367b2de84c34435d17d65125e5b2d31704f5934863ad5a43deef62cb3cc8314b24ab1d33328bf036a735d08022198960f3a7bb7da1c740cdd082addb78423162e00dc3d46bec006d0d9bd972d1f2822351e8c3b4ba0a4b34efceb2aa07e27d9ee19e861ab87cc4a4ab0ccf1842053e0fc11d9e792fcf642f7c22f6cfa470c83c7eea66db066bb99bb0e313c6419a7d3a6933d650d2075b553cd99d4d1a773eb3c6611b013a7444a01d89bb59b76c92523efbab92399fd1c1820b0a46e413f5fdb7553776252543cea685e31adab2a2499149e6e8bf7779924d81c2714aab9e2d10995a07220fc8b69bd8188819356329a74fc782baeec84f9e5cfa7325fb47008f57d04b10d652bc1b5e317dfa71a4c379b07e636bdcb7a944920ddb4523a50d1f0b8e8735cc3ceb33c6c8c42888040e8e909dc8593edad2d6eed931a9881e662c72f43f837ad0274f8845268d544ac59a8d797075c761fcf4bbc892d4fbe031c70097f82157abae5eca1b0559828e422564a3636e6621cf40d0a59daf6738cca3c5ef1e1c51d9f2711335b6946a709c0fb131f4dbd8f89dd2e72d4a4eda83f6fb84f86569a9142a60f26757833aec49f015f18e73f142ea790af766bddbff3a3337ba6fef219834a597a6651bdac2e741e75ace4da1a934fc28e183a497682fc80ecc0e41a43edadcfab9d428ff10088abbec78889e855bad736bc0480d57231d5672ebc55ded6b4aba8a2fda397351f5ffe7d448b1ba55439d6fcc51110b76ce34e27fdaca5bbcd28bbda386ba47853a9e978a93aa3578babf398e67f30e47f18abdd9e4dfc89d66a8017a26417ae83440098ba749ffde54aba8ccd359202b98f6ca041f19b1df6b239d572202a4ca6fe98980bf49c8d3e39b7ce98f0730113720bf52b71005fea87105dd0364b004f734b669e6dad130327385c706c5004eddcbbda177b272319b37e32d82e873b6b5ec56972fa86c1f3555d3f62d8ef5cbc45ef64c31a21c819f16d1a0c985e080541f5fdb610efb5e2ff8dc17206b827feb646093b2aac95f1bad08ee3a0fbae40749c34da4188909923adf6b273437a102167c4a22d7b11a7c5df6876590261564825cdfde09cd87a1086cab90c786d650bced7e5f060de4e78007d7de9015a5887aba4b51be34b945f2f75d36562b9e8042155a8138f6a93d561fa4fb33b8cac57fa6627ba93ddf540ddf95c5430b1d539113f844c22f5378a1c083c093143918ffc0758f2ebdb7ecac11858571d9a3a03c833a7a834bc556409b80e295a83f825c728b72f59a12e68d840271544ac7bd48c40f8291ae2d96458eb7e59f720a8e073be1558d0b942438dab7f9007a6f1ed4ce55ce07ed28055383143ac7a61831c10d01c04530373f2b4cfe478f0bafc7d49dba637986c9af966a957eef5c15e203d192bca08521d7d4a234e0ea4ba1605ea7a98c4756fc06813ad71c5932583f61dae91f512a74ab1b4eeebb88fbbe9256d041fe8ea7ff15d757d9e07e6d0cbd38bd0cfe851bc60a616e6e7feec9d14e263e69b6e499096dd28fe1655f6032fc39678ffcd8002b1146dc4af90be09ba1f0df5effe172d79e5eca8b519dfc3585a326bb1a8e970fe103b5e44bb2e9b217713e8d02a9131ddb3bb323139423c62ae32224fb6da9894c397e4a4d88868b242b08ac7643ed9c381ed83c1a131c7d53cee7da94064f2c2866652827310f96d52a6bf341f9374e5b48bb3fa520abee80cb2d1906a8cd5ef070255297004dc389d2d4739d0f5741d234aa595c7cea5274d732f0c562e14e4db08024380b2bf131bba75682dd378d136e2e8a1a29a521af5f9631a14d09c5d9965b9e7d3151d0ca3f638ba6c16de16762d66ac2dd9d27abd588d98af348ee53dc787a7bd16e9110adfe2ea2f203ce15eed545393f0e7d824ad925c7ee4000bb3fc3c679b13702cb789e27428fcf6af0462799303d24b2b4dc5212ed356933c52077eac0927b4c9580e78cd2fadfcd05789b664ec40ddc92cbc19c2911610a5d31eaa29ea8ffaaa63b72cc01b4d78657c1efb23a711f072e5708938f2270f6e9e35c0de97a3155ac7b855110ffb6eac717ba24b8edffa4a2970b9bf9bc8a84ce3959341a921f6477a4004b90797a97c6eb3af5b73a3c5c777388bb667f8f00829df7032c42e8a279cd2986b9f92e997c3fa9ea2d299120ba7b0e9e3951c50c55db9570899b0076f157c2242602b83c2736951ed2381bc74ef84b52de3736e60c92bc978dca9155c5c2f25d409a5994d1d9df95d580e163baa4cd276434296f0a3e5286d9ceeaa1d8834284366b7b10b80c341a0d24608c6a09e3e09b03db789e3f8157e4af90bf581b833b1fae12c63335ef58787600577e2fd6f7363dfdd79fd460abcc98792a40e84f2000a0d77c91e95cbccc671e77d06a69d27e682cb4b50e2f94d6f5f41ff67cda3b1da86efe0c0f9fe38b247ca7f1a8f4e6c44aaed96679b7a1a20567e07772247876f21d598b820606f73c0860bf7410194675fc65b756dd7ffa9b040dbde30678cdf074b8f679739d88dffb95f010502502c9bc5ddae593a1b784efe741561a4b744da51505d7b77ae79f825863eef160c2954e4f081f282c404f515b7379869ba0aab2b9beccd0d7dff6151e243039565f8690c2cfe2e6e9f7042f3b4f5f71788182a0a5acb2c6cbcecff510181c4a57646586a0b4ce000000000000000000000000000016253742';
-const CHALLENGE_MCU = '7fd384fa2d06a8bcfd796b2b05a9c7d03a545199e8ed30c30e19d559a71796e4';
-const SIGNATURE_MCU =
+export const SIGNATURE_MCU =
     'e22777813b2b8e6570292d7effd8996369a84f2242aa0cd7589f4a6dfde6d7067a2e890d5d517cf134caafd272358e6ee6d9674eb2896d58f1a487f77e00ed4877b3469ffb0a0ddd37e6b3405009079fe274c53f30991150ed7d0d87d9de1664d1a2c93e85a74ef83dcba35789af3eec1191b61f703b735c37892419e6de7cee178abec9bade984fd5bdbe28b3271bba2cb0c7615e28eaefa07661cb97503af7eac8aba4a42d17c858dcad3ab519c1d32c3c8bd15dbba14a51382181344dd79cf0b116e695abd804f9caceb202b025b0591f87567cfb15a9245d339e1185663882efb2adcc8fa22225d50ea4cc244af916a4294455330797dc62ca4d6d794ae7850ac3fd5e8f1508c07ac6fbfedf9ee49c279c672b978d9e4b4a653e330e663973ab0181e6eb5f505929c84a8ed40125f0d11e0eb1d0c3c9fc1cdc75f8841163003419fabea072484fac34201378cea781baa3d82781600ac4c0f715c189c434ccdb135f91965bf17af018455a111fc02f2dd60d8c0bf1fef698ee0d126df94d1eaf4b1722a7a5f14e368126f9177cf46f7c257b4860cb55a6134332f04e905ff52bfc193d81a4bc751ec8e309e86078d631c9b4050c027d14743ae6833fdaf8bda6ba62a60b4b1a629cf78674eaee15d19c7dbaebc6266e949cf7a076e91708371e4d96b2ea4156a2c3b2f19f81e68c37e1e610a99e2374980e707190377271f7c18fee01b06fda52d36add90e6ed7793810597c437751c3c36d6f9c630a4791fef43e5a7280c053f5b8c254b7df2d811dbbd665c244987cbae7b185497c62bf831b2bd0030bb2bd79343cc2ce2d3a46325bbac1a77c98e6fdb44ac9618de6c6c527b24c97a7d0130e138fe457b5e9d6aa2d10d90443ff13168149db7b5dd744af81cca7de57664330ab277f933e4de40986f503fa6d9abc68933ef73edb0d56fa5750f41d7871fbea4a8c9593437bd875084f9c6b8721f7778d643a5c168234b64e67a9824724959ea2dc337f026ee756f596b8295b9bcbbafebe6726051b52fc0395130d73d5e3016d18f61b4ccbf00c69c06f2d8ed097a1da8a6e785430100fe5d5b2b154bf6220132aeda1574679446fc9e25c28fba02b5105b47f9f78907d708d6d2074d95505ea09408faa8010ab02dc72f5da65eb6445d63832bfff25861b30db00cfb49cfa35be515c64cb33b1d9bc16736987e36430b517f512b35dc4755f5e595e5e2e35f06c9d5ff9d7e0f9783ded0680a7a89e831099155a551161611e036a9f0585848667fb2c6d5f108079a32f0f9adb86a37c39c7cb9ca4cdbf4347169f683f729e5c95381891ad5fcf1a5474b3a5b69e8e30a6edbe2ed3fe5f18d4c8e98e66f85c56660ff3326e4f7af082767bb3b1620e197474c0d261e92a6d4f108afb942e5946ecb858e7809b3164faffd05d1cdab3d01b9dc9afb9034bf353c5fe8a3fff649cbed0b1c28c69187765a1af67cfad96fb190538551d5e938334181be1061405e07d1b6e353d1d4485e1e66ca2c8ccf8440f9c6c718f7744c367e48e179f4f99ae1c8473e778a739b6e25981fa55c47b42904ac1cab12ef62e95b752586de89458ff9c376514b8be9b16c3adf6fbc94db09206e3875a7ae19df3d52178dc1d41429f75834bc84a7116a48239d13876a8dfb05e9068eb521cfdcb62025b90a2353268f455b90957eb1c54e4878b03fd0277da65f60665e66602f0041c3e278559926c1ec8895d277e81f36b4141c71832c5dcb223d683909e6a12d745d593dfd75f6529d028917f8131f6cf89b7410e740d436ddeface0c0fe9b05c55d6f8d14d408039bd9f373b3c27e826e33c806317537751e9898d4c00ec2d8098c47dc7731d7e7d86ce28d598cd1dce370965088949906dd7941ce0f7eedae10f2d952dca65d0933f2c442b1e199142467ed9ca09f1c040ddf9cd485b895e594a9ef54f8ea8b9c0b682975ce9a4229d8a2832b74abbcd6c854a8c2e266aa8780c04f650ae2e7523f982e19165e24247e9e6a1d801b6a2a0fc8e377c47e368c1e0360f29a3975b5e35caf0591b38f2b2c35e565fb8793c0f13e87738f0637ffcb54b204230a5c3ff06b2082f4da3aa818795c7682aa44a77137bc7ff77964f5699854dbcf15be421b351455f489fec312a9e60acf4b5b281d29e19713c1a38f38b26125a9440d89a0a14597e888fd71ff5eb172b00ef2e26a68b2dfe963c1ffb99e83cb8eb4187fb5a671910ac2602bf09ee4640c18577171db843e59f82a506313d8b5428c30decb31768569a45119f5142809f8804c58459376f5026ff30b94060c20bc9cc44fe8560d886e10b751ce7277a6307a05175bad09dcf0db0e19e2f4c065c2ecc7f69ef54bc2b27e1bfed01e9e6ecbd6aca01583877262252a1d3c80a4b5afc25a598e631513063b201512a62bdd39e4c209b93b049f84b40ae9be18fe050352eee5158d2fabe0f01791e57ca16616936b17439eb7a0977e99967d01a24b4f3235a52e0b3bf96b6b0692840e996fc6384af8f491b4b872d9bcebf7af2dc32130dddcea63bee5d4ccf955237f39b2f93719c2c835deea23fbd6ffd529c9dae92cf4aa2d389f7a95b5a98d4536195a5107b73fd5b3dd6d9e6f8d22b2ea62064dbe83f9e4951ffd32b9ad4674135a924219cc0971085dd9437ed3d91f2fd466975c0c2d4426bf206e8acf73386927ce6e39b8a75e09aa548ce3f66151e2297c9741bed7fbc5519653d64566c30c4bca023a6eafb450b5a70cacfef4bf374ce85adcd7f0d0270d668f4ffd3a32521360f77f55470705fc24ed702bf180e1515e008568c978114333ee3eb42fe06a4b4357150f347dd8bd6946325369e39898f30614803dd6492f0a6a13692adca9ec6229e37edc57c775d0345b0168b1e6d24c42c6b382c6a0fe3040caf3142126cb965fa5aa052de4e95bd16cecc6056ac93c961650a4d8168228f4b94c806b08f72e0500697a52ca89ce0c4beda6839c24425c4aafe4a0e59f0f57db078c01f8ab3f02258823dc11ef55a973fcb432024dbc8d498e98fe775fb85c1c707403840e8279018e12836fcc26f53ccaf4bd2345f4509a8d0a1579ae758611a06159dd5f93fb800dd93ba22ee8fab9052630b96e1396fc27fefaf44a916548e9188b33c4c8e95c4b2615748c19ceb96f1bc2c5fcc486ceff56e673803bedd0e14c9faf56a74aaff4f04c06ae4e73736d94bcc0151d434286f1c4e7412185efb2ba8ff43aca7261dab5b012b9643eb8029cef0dac399b95d72e57da4796ebd93f7d1c5a66b391691e56956d6c57080c1223415889c1c4c6d9ebf5172e444c697887898a91a0afb4bdc0c4def3fb0003070c222b3e51585c5d6873797a80889da7b6d5dcf2fd151c3b4647494f6378969da0bdd8e1e6e7eaee00000000000d20384b';
 
-// This one of dev pubKeys, but not the one published as DEBUG key.
-const T2B1rootPubKeyOptiga =
+// This is one of the dev pubKeys used internally, but not the one published as the DEBUG key in deviceAuthenticityConfig.
+export const T2B1rootPubKeyOptiga =
     '04626d58aca84f0fcb52ea63f0eb08de1067b8d406574a715d5e7928f4b67f113a00fb5c5918e74d2327311946c446b242c20fe7347482999bdc1e229b94e27d96';
 // Same as deviceAuthenticityConfig.T3W1.debug!.rootPubKeysTropic![0]
-const T3W1rootPubKeyTropic = 'cd318dc8405ae4f4144e3284dcb7b0cb0f0c2195c2ca14a0f6fccd9104e32a4b';
+export const T3W1rootPubKeyTropic =
+    'cd318dc8405ae4f4144e3284dcb7b0cb0f0c2195c2ca14a0f6fccd9104e32a4b';
 // Same as deviceAuthenticityConfig.T3W1.debug!.rootPubKeysMLDSA![0]
-const T3W1rootPubKeyMLDSA =
+export const T3W1rootPubKeyMLDSA =
     '1f3e5edad45fc6980dbfd597ed17b77e4d4ad1b1ec4804e04dc061cc2cfe77bc96fbc93f698f1fe00c47177c8150276dab86cbf407add7527deb09b65a6623ef581422292d77744be64f4401ca65b9bd027e32a76f307413b086869eccbcfc5e8823be8e0b3897727bca568feeff62957cd1e261e4aab67271725d240c7e7d2a20a5c201699be83d6c6a57fdfc09ac9b6b599e2a14a50e421d246d1d1df805db0b25534e500fed76fa89975a4270e274a9fa72b528e47d999abdeca5e08e51d447ae4019f699088d485bb52d5eec7a8c725cd34ff955c16437877a131f068438b5699c2285000463e24c633b023cde279be015db2f8b16de07fa82ee9adcbf868b898e25eeef5d60d03e1da874aca6f2089b1e7c72d7b532d3a27e38936d53c0080f9908b6727b13fe961ec8b08004f077b9cce95b5eee9b44d351e79eb222aba1bc18895d36cf4843ffcdfc735e1b66f3e72db5aaf8b5ecf56dbf0d616cf53d192a60e61ad3c51a595213100fd00835434a28aab3eb5949d3405de5db1a586bd6e937444913e9f1d4c00247acd49a2ed9953636e5d74417b4c7144c6c22fd8f7e25cb6cf6c8510de6915b74c95a28c197152208096d0249b86150e807978ca985c110d07b7d0a24434039f6ef12a69d3e073af6ae44e7c4606a354a3e953bdcff10c8dc68b807f41ac563ff0d7141c328b34152a25a16e9cb8228f7707356d2d7f3caa8d778a85e3b66242c124281c0f3201ad28596b868db138149eb98b821dcaaf86b5e11c2ac1161975bf0b2ecefc111b7b7401ce261911526618044d126eeee885efe51be0e32e93aaaffe34899df5b10df7c5a9c54c5b4704671640a92e8631f0d401f44c2238b5f9a37499b3d83c1c60096f2b48c3081dd9769db325c1cd1ffe84bfd8a80dde3cd650c49e14af3ab6c317ee74bd39e976ff5b7ad490e80f3eff9edf3b3855070a5a06fbe9cbca54f521e3d74cf13a4d116e23e531a7c70232c044b7e7d02467afe2a40def14298688eced5d3d87a102d31d7cd5409918ced4bab190505df48f451deea1c29168246d621529f0346eaab0e2302f3ce206dd84837704b8bdedbadea5b5822cdecbfdb180cde14e62ee26a929f0a37a99bacfe237376a3e6c289b06b8e93ed9d1435e3ee6399ab1a06d4f70097f80e26fe6f60b27a68104f6333b8db1bc78e12d54647b23feb0f11ae070f3a6be9d9f4d89a6b5ef8595d010d410e72138cfcbd5e41cca1df33ddad657b699cf4d3dc602888e5aaa0a7a2325ba4a36d7de49c591d8e355fd1de644d772bf2a060093743388dbe5d30a2d97304abc17c547bd65276291d6aa7b0723703733239fabef39c2de47071d50074e68fbe71c88f75d4510bb5f6bc244b3e7b219eb89fe5f148459a2d693cd3fac32aa069cbdd4e27c9778096f6805bf6df9aec9671f0a58b28704f74775e81ac46826c1c4709b0e0989d6e8e73bd1795ce85d85c14ad2997cd2810272a5682e7ed9aa7a092708c0f4bc67f68d317d406a96ddd2842ea22cbf3410cd7a77005b61115b794894cf9d10ebb4f5552bec8c21d90ba97d6139d5261d7f51f4009c3618f9feb22cc929ade204eca7c0f813409ceb88a13512f85e03948314ba9c4488aafda88ce447a0e052a6a605e32d04a20e2e52c41f371247963f24fc28c5cc1ddd8fe1681a2a293daf2688e9b026303a76cc60f40290c4d823efb85f36cb5e092369cc20515a8077e7702ed54ca1565d37b83184ee46c4a6adf2840f5a5fb002db70690bbf93888e685fdd1f9ea211b42693e0bd5ba5235a6ea1b714774d13279d7db3440f7fbced53f203ce6500555bcb9251e6f36d3352326501c';
 
 const T2B1_CONFIG: CertPubKeys = { rootPubKeysOptiga: [T2B1rootPubKeyOptiga] };
@@ -54,7 +51,7 @@ const T3W1_CONFIG: CertPubKeys = {
     rootPubKeysMLDSA: [T3W1rootPubKeyMLDSA],
 };
 
-const CONFIG: DeviceAuthenticityConfig = {
+export const CONFIG: DeviceAuthenticityConfig = {
     version: 1,
     T2B1: T2B1_CONFIG,
     T3B1: { rootPubKeysOptiga: [] },
@@ -62,7 +59,7 @@ const CONFIG: DeviceAuthenticityConfig = {
     T3W1: T3W1_CONFIG,
 };
 
-const CONFIG_WITH_DEBUG_KEYS: DeviceAuthenticityConfig = {
+export const CONFIG_WITH_DEBUG_KEYS: DeviceAuthenticityConfig = {
     ...CONFIG,
     T2B1: { rootPubKeysOptiga: [], debug: T2B1_CONFIG },
     T3W1: {
@@ -73,7 +70,7 @@ const CONFIG_WITH_DEBUG_KEYS: DeviceAuthenticityConfig = {
     },
 };
 
-const BLACKLIST_CONFIG: DeviceAuthenticityBlacklistConfig = {
+export const BLACKLIST_CONFIG: DeviceAuthenticityBlacklistConfig = {
     version: 1,
     blacklistedCaPubKeys: [],
     debug: {
@@ -81,7 +78,7 @@ const BLACKLIST_CONFIG: DeviceAuthenticityBlacklistConfig = {
     },
 };
 
-const defaultOptigaProps: VerifyAuthenticityProofParams = {
+export const defaultOptigaProps: VerifyAuthenticityProofParams = {
     certificates: [DEVICE_CERT_OPTIGA, CA_CERT_OPTIGA],
     signature: SIGNATURE_OPTIGA,
     signedData: prepareDeviceAuthenticityData({ payload: Buffer.from(CHALLENGE, 'hex') }),
@@ -89,334 +86,17 @@ const defaultOptigaProps: VerifyAuthenticityProofParams = {
     config: CONFIG,
     blacklistConfig: BLACKLIST_CONFIG,
 };
-const defaultTropicProps: VerifyAuthenticityProofParams = {
+
+export const defaultTropicProps: VerifyAuthenticityProofParams = {
     ...defaultOptigaProps,
     certificates: [DEVICE_CERT_TROPIC, CA_CERT_TROPIC],
     signature: SIGNATURE_TROPIC,
     deviceModel: 'T3W1',
 };
-const defaultMCUProps: VerifyAuthenticityProofParams = {
+
+export const defaultMCUProps: VerifyAuthenticityProofParams = {
     ...defaultTropicProps,
     certificates: [DEVICE_CERT_MCU],
     signedData: prepareDeviceAuthenticityData({ payload: Buffer.from(CHALLENGE_MCU, 'hex') }),
     signature: SIGNATURE_MCU,
 };
-
-// TODO rewrite tests its with a fixtures object to deduplicate & to test the atomic functions with the same data.
-describe(`firmware/${verifyAuthenticityProof.name}`, () => {
-    it('verify success for optiga (with prod keys)', async () => {
-        const verify = await verifyAuthenticityProof(defaultOptigaProps);
-        expect(verify.valid).toBe(true);
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(T2B1rootPubKeyOptiga);
-    });
-
-    it('verify success for tropic (with prod keys)', async () => {
-        const verify = await verifyAuthenticityProof(defaultTropicProps);
-        expect(verify.valid).toBe(true);
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(T3W1rootPubKeyTropic);
-    });
-
-    it('verify success for MCU (with prod keys)', async () => {
-        const verify = await verifyAuthenticityProof(defaultMCUProps);
-        expect(verify.valid).toBe(true);
-        expect(verify.rootPubKey).toBe(T3W1rootPubKeyMLDSA);
-    });
-
-    it('verify success for optiga (with debug keys)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultOptigaProps,
-            config: CONFIG_WITH_DEBUG_KEYS,
-            allowDebugKeys: true,
-        });
-        expect(verify.valid).toBe(true);
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(T2B1rootPubKeyOptiga);
-    });
-
-    it('verify success for tropic (with debug keys)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultTropicProps,
-            config: CONFIG_WITH_DEBUG_KEYS,
-            allowDebugKeys: true,
-        });
-        expect(verify.valid).toBe(true);
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(T3W1rootPubKeyTropic);
-    });
-
-    it('verify success for MCU (with debug keys)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultMCUProps,
-            config: CONFIG_WITH_DEBUG_KEYS,
-            allowDebugKeys: true,
-        });
-        expect(verify.valid).toBe(true);
-        expect(verify.rootPubKey).toBe(T3W1rootPubKeyMLDSA);
-    });
-
-    it('verify errored for optiga (debug keys not allowed)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultOptigaProps,
-            config: CONFIG_WITH_DEBUG_KEYS,
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for tropic (debug keys not allowed)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultTropicProps,
-            config: CONFIG_WITH_DEBUG_KEYS,
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for MCU (debug keys not allowed)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultMCUProps,
-            config: CONFIG_WITH_DEBUG_KEYS,
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for optiga (signature mismatch)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultOptigaProps,
-            signature: `aa${SIGNATURE_OPTIGA.slice(2)}`, // invalid 1st byte of signature
-        });
-
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('INVALID_DEVICE_SIGNATURE');
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(T2B1rootPubKeyOptiga);
-    });
-
-    it('verify errored for tropic (signature mismatch)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultTropicProps,
-            signature: `aa${SIGNATURE_TROPIC.slice(2)}`, // invalid 1st byte of signature
-        });
-
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('INVALID_DEVICE_SIGNATURE');
-        expect(verify.caPubKey).toEqual(expect.any(String));
-        expect(verify.rootPubKey).toBe(T3W1rootPubKeyTropic);
-    });
-
-    it('verify errored for MCU (signature mismatch)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultMCUProps,
-            signature: `aa${SIGNATURE_MCU.slice(2)}`, // invalid 1st byte of signature
-        });
-
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('INVALID_DEVICE_SIGNATURE');
-        expect(verify.rootPubKey).toBe(T3W1rootPubKeyMLDSA);
-    });
-
-    it('verify errored for optiga (missing rootPubKey)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultOptigaProps,
-            config: {
-                ...CONFIG,
-                T2B1: {
-                    ...CONFIG.T2B1,
-                    rootPubKeysOptiga: [
-                        // invalid root pub key
-                        '0423a5c9ec44dfb96023838d958f6289fa611277ee7af60c3bcb54eff2310546d5ece48b1a507503142b122b53eda53fef3f3d3510f3b7ae2fd5f13614b025ede1',
-                    ],
-                },
-            },
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for tropic (missing rootPubKey)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultTropicProps,
-            config: {
-                ...CONFIG,
-                T3W1: {
-                    ...CONFIG.T3W1,
-                    rootPubKeysTropic: ['deadbeef'.repeat(8)],
-                },
-            },
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for MCU (missing rootPubKey)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultMCUProps,
-            config: {
-                ...CONFIG,
-                T3W1: {
-                    ...CONFIG.T3W1,
-                    rootPubKeysMLDSA: ['deadbeef'.repeat(328)],
-                },
-            },
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for tropic (no rootPubKeysTropic defined)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultTropicProps,
-            config: { ...CONFIG, T3W1: { rootPubKeysOptiga: [] } },
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for MCU (no rootPubKeysTropic defined)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultMCUProps,
-            config: { ...CONFIG, T3W1: { rootPubKeysOptiga: [] } },
-        });
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('ROOT_PUBKEY_NOT_FOUND');
-        expect(verify.rootPubKey).toBe(undefined);
-    });
-
-    it('verify errored for optiga (device model mismatch)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultOptigaProps,
-            certificates: [
-                '3082019e30820145a00302010202044ee2a50f300a06082a8648ce3d040302304f310b300906035504061302435a311e301c060355040a0c155472657a6f7220436f6d70616e7920732e722e6f2e3120301e06035504030c175472657a6f72204d616e75666163747572696e67204341301e170d3232303433303134313630315a170d3432303433303134313630315a301d311b301906035504030c1254324232205472657a6f72205361666520333059301306072a8648ce3d020106082a8648ce3d030107034200049bbf06dad9ab5905e05471ce16d5222c89c2caa39f26267ac0747129885fbd441bcc7fa84de120a36755daf30a6f47e8c0d4bddc15036ed2a3447dfa7a1d3e88a341303f300e0603551d0f0101ff040403020080300c0603551d130101ff04023000301f0603551d23041830168014176d8b9a403574f6a2b9ac353ef578682201a21a300a06082a8648ce3d04030203470030440220747c545e112df816173d3071f1ab25d399d8108550764ce1a3a428f1f18b506902200cda822c75b3da6e44e098014452f3fc324f29a79204c3fb4d5815afafc04b17',
-                CA_CERT_OPTIGA,
-            ],
-        });
-
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('INVALID_DEVICE_MODEL');
-        expect(verify.rootPubKey).toBe(T2B1rootPubKeyOptiga);
-    });
-
-    // device model mismatch case is not relevant for tropic or MCU (supported only by T3W1)
-
-    it('verify errored for optiga (caPubKey on blacklist)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultOptigaProps,
-            blacklistConfig: {
-                ...BLACKLIST_CONFIG,
-                blacklistedCaPubKeys: [
-                    '041b36cc98d5e3d1a20677aaf26254ef3756f27c9d63080c93ad3e7d39d3ad23bf00497b924789bc8e3f87834994e16780ad4eae7e75db1f03835ca64363e980b4',
-                ],
-            },
-        });
-
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('CA_PUBKEY_BLACKLISTED');
-        expect(verify.rootPubKey).toBe(T2B1rootPubKeyOptiga);
-    });
-
-    it('verify errored for tropic (caPubKey on blacklist)', async () => {
-        const verify = await verifyAuthenticityProof({
-            ...defaultTropicProps,
-            blacklistConfig: {
-                ...BLACKLIST_CONFIG,
-                blacklistedCaPubKeys: [
-                    '9603b4971f811ed2a1cdb9ec3e6d6d0e22facfd83892a30480460872a2003f45',
-                ],
-            },
-        });
-
-        expect(verify.valid).toBe(false);
-        expect(verify.error).toBe('CA_PUBKEY_BLACKLISTED');
-        expect(verify.rootPubKey).toBe(T3W1rootPubKeyTropic);
-    });
-
-    // caPubKey blacklist not relevant for MCU as there is no CA certificate
-
-    it('verify signature for evolu (optiga)', async () => {
-        const header = 'EvoluSignRegistrationRequestV1:';
-        const challenge = '6006b5c68105bd4872129e3ee42383e1f9404c711e433b88557c22c3f94d357e';
-        const publicKey =
-            '0487472eec47aa28fa62ff3231f60b5c89751318a5598af5f93ab2aad9061ca25f53b352a97b855f16b11b795b715249c8dbfb6e47339f677e30d530f0e80bc4bb';
-        const size = 1000;
-
-        const sizeBuffer = Buffer.alloc(4);
-        sizeBuffer.writeUInt32BE(size, 0);
-
-        const bufferChunks = [
-            Buffer.from(publicKey, 'hex'),
-            Buffer.from(challenge, 'hex'),
-            sizeBuffer,
-        ];
-        const signedData = prepareDeviceAuthenticityData({ payload: bufferChunks, prefix: header });
-
-        const verify = await verifyAuthenticityProof({
-            ...defaultOptigaProps,
-            signedData,
-            signature: SIGNATURE_OPTIGA_EVOLU,
-        });
-
-        expect(verify.valid).toBe(true);
-    });
-});
-
-describe(`firmware/${matchRootPubKeyToCertificate.name}`, () => {
-    it('matches the valid Optiga root public key for a CA certificate', async () => {
-        const cert = parseCertificate(new Uint8Array(Buffer.from(CA_CERT_OPTIGA, 'hex')));
-        await expect(
-            matchRootPubKeyToCertificate({
-                allRootPubKeys: [T2B1rootPubKeyOptiga, T3W1rootPubKeyTropic],
-                cert,
-            }),
-        ).resolves.toBe(T2B1rootPubKeyOptiga);
-    });
-
-    it('does not match any root public key for a device certificate', async () => {
-        const cert = parseCertificate(new Uint8Array(Buffer.from(DEVICE_CERT_OPTIGA, 'hex')));
-        await expect(
-            matchRootPubKeyToCertificate({
-                allRootPubKeys: [T2B1rootPubKeyOptiga, T3W1rootPubKeyTropic],
-                cert,
-            }),
-        ).resolves.toBe(undefined);
-    });
-});
-
-describe(`firmware/${prepareDeviceAuthenticityData.name}`, () => {
-    it('prepare data with single buffer with default prefix', () => {
-        const payload = Buffer.from('deadbeef', 'hex');
-        const data = prepareDeviceAuthenticityData({ payload });
-        const expectedResult = Buffer.concat([
-            bufferUtils.getChunkSize(19),
-            Buffer.from('AuthenticateDevice:'),
-            bufferUtils.getChunkSize(4),
-            payload,
-        ]);
-        expect(data).toEqual(expectedResult);
-    });
-
-    it('prepare data with multiple buffers with custom prefix', () => {
-        const payload = [Buffer.from('dead', 'hex'), Buffer.from('beef', 'hex')];
-        const data = prepareDeviceAuthenticityData({ payload, prefix: 'Something:' });
-        const expectedResult = Buffer.concat([
-            bufferUtils.getChunkSize(10),
-            Buffer.from('Something:'),
-            bufferUtils.getChunkSize(2),
-            Buffer.from('dead', 'hex'),
-            bufferUtils.getChunkSize(2),
-            Buffer.from('beef', 'hex'),
-        ]);
-        expect(data).toEqual(expectedResult);
-    });
-});
