+++ packages/device-authenticity/src/__fixtures__/verifyAuthenticityProof.ts
@@ -1,0 +1,3 @@
+type Fixture = {};
+
+export const fixtures: Fixture[] = [];
